(test) update tests so that each component's tests are in their own file
