+++ test/client/appSpec.js
@@ -6,9 +6,6 @@ import App from '../../app/components/App';
 import Header from '../../app/components/Header';
 import NewsList from '../../app/components/NewsList';
 import Topics from '../../app/components/Topics';
-import TopicsList from '../../app/components/TopicsList';
-import TopicsListItem from '../../app/components/TopicsListItem';
-import TopicsSearch from '../../app/components/TopicsSearch';
 
 Enzyme.configure({ adapter: new Adapter() });
 
@@ -53,47 +50,3 @@ describe('<App />', function () {
     expect(wrapper.find(Topics).length).toEqual(1);
   });
 });
-
-describe('<Topics />', function () {
-  it('contains a <TopicsList /> component', function() {
-    const wrapper = mount(<Topics />);
-    expect(wrapper.find(TopicsList).length).toEqual(1);
-  });
-
-  it('contains a <TopicsSearch /> component', function() {
-    const wrapper = mount(<Topics />);
-    expect(wrapper.find(TopicsSearch).length).toEqual(1);
-  });
-});
-
-describe('<TopicsList />', function () {
-  it('contains a <TopicsListItem /> component', function() {
-    const topics = ['politics'];
-    const wrapper = mount(<TopicsList topics={topics} />);
-    expect(wrapper.find(TopicsListItem).length).toEqual(1);
-  });
-
-  it('doesn\'t render <TopicsListItem /> component if no topics are passed in', function() {
-    const wrapper = mount(<TopicsList />);
-    expect(wrapper.find(TopicsListItem).length).toEqual(0);
-  });
-
-  it('contains a <TopicsListItem /> component that dynamically renders topics', function() {
-    const topics = ['politics', 'art'];
-    const wrapper = mount(<TopicsList topics={topics} />);
-    expect(wrapper.find(TopicsListItem).length).toEqual(2);
-  });
-});
-
-describe('<TopicsSearch />', function () {
-  it('contains an onSearch function', function() {
-    const wrapper = mount(<TopicsSearch />);
-    expect(wrapper.props().onSearch).toBe.defined;
-  });
-
-  it('contains an onSearch function', function() {
-    const wrapper = mount(<TopicsSearch />);
-    expect(wrapper.props().handleBarChange).toBe.defined;
-  });
-});
-
